app/app.py
import streamlit as st
from PIL import Image
import numpy as np
import torch

# Placeholder for the model loading function
def load_model(model_path):
    # TODO: Load your trained PyTorch model
    print(f"Loading model from {model_path}")
    return None

# Placeholder for the prediction function
def predict(model, image):
    # TODO: Preprocess the image and get a prediction from the model
    # Returning a dummy prediction
    return "Normal", 0.95

st.title("Chest X-Ray Disease Detection")
st.write("Upload a chest X-ray image to classify it as Normal, Pneumonia, or Tuberculosis.")

uploaded_file = st.file_uploader("Choose an X-ray image...", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption='Uploaded X-ray', use_column_width=True)
    st.write("")
    st.write("Classifying...")

    # model = load_model('models/your_model.pt') # Uncomment when model is ready
    # if model:
    #     label, confidence = predict(model, image)
    #     st.write(f"Prediction: **{label}** with {confidence:.2f} confidence")
    # else:
    #     st.write("Model not loaded.")

    # Using dummy prediction for now
    label, confidence = "Normal", 0.95 
    st.write(f"Prediction: **{label}** with {confidence:.2f} confidence")

    # Placeholder for Grad-CAM visualization
    st.write("### Grad-CAM Explanation")
    st.write("_Grad-CAM visualization will be shown here._")
