utils/grad_cam_utils.py
import torch
from pytorch_grad_cam import GradCAM
from pytorch_grad_cam.utils.image import show_cam_on_image
from pytorch_grad_cam.utils.model_targets import ClassifierOutputTarget

def generate_grad_cam(model, image_tensor, target_layer):
    """
    Generates a Grad-CAM visualization for a given model and image.
    """
    # TODO: Complete the implementation for Grad-CAM visualization
    print("Grad-CAM generation logic needs to be implemented.")
    return None

if __name__ == '__main__':
    # This is a placeholder for testing the Grad-CAM utility
    pass 